[WIP] dynamic import for lite
